api/requirements/development.txt
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
django-debug-toolbar==4.2.0
ipython==8.18.1
black==23.12.1
flake8==6.1.0
isort==5.13.2